requirements_ml_optimization.txt
# ML Optimization Dependencies
# ML优化依赖包

# Core ML libraries
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Hyperparameter optimization
scikit-optimize>=0.9.0

# Model interpretation
shap>=0.42.0

# Property-based testing
hypothesis>=6.82.0

# Configuration
pyyaml>=6.0

# Utilities
joblib>=1.3.0
